rename logs packages
