+++ pkg/logtailing/tailer_test.go
@@ -1,4 +1,4 @@
-package logs
+package logtailing
 
 import (
 	"fmt"
@@ -8,12 +8,12 @@ import (
 )
 
 func TestJsonifyFiltersAll(t *testing.T) {
-	filters :=&LogFilters{
-		FilterIPAddress: "my-ip-address",
-		FilterHTTPMethod: "my-http-method",
-		FilterRequestPath: "my-request-path",
-		FilterSource: "my-source",
-		FilterStatusCode: "my-status-code",
+	filters := &LogFilters{
+		FilterIPAddress:      "my-ip-address",
+		FilterHTTPMethod:     "my-http-method",
+		FilterRequestPath:    "my-request-path",
+		FilterSource:         "my-source",
+		FilterStatusCode:     "my-status-code",
 		FilterStatusCodeType: "my-status-code-type",
 	}
 	expected := fmt.Sprintf(`{"filter_ip_address":"my-ip-address","filter_http_method":"my-http-method","filter_request_path":"my-request-path","filter_source":"my-source","filter_status_code":"my-status-code","filter_status_code_type":"my-status-code-type"}`)
@@ -23,7 +23,7 @@ func TestJsonifyFiltersAll(t *testing.T) {
 }
 
 func TestJsonifyFiltersSome(t *testing.T) {
-	filters :=&LogFilters{
+	filters := &LogFilters{
 		FilterHTTPMethod: "my-http-method",
 		FilterStatusCode: "my-status-code",
 	}
@@ -34,7 +34,7 @@ func TestJsonifyFiltersSome(t *testing.T) {
 }
 
 func TestJsonifyFiltersEmpty(t *testing.T) {
-	filters :=&LogFilters{}
+	filters := &LogFilters{}
 	filtersStr, err := jsonifyFilters(filters)
 	assert.Nil(t, err)
 	assert.Equal(t, "{}", filtersStr)
